Pages: UI improvements, hover over 'deleted' to reveal summary
Check for recreated pages and label them

Various bug fixes

Bug: https://phabricator.wikimedia.org/T165864
