+++ app/Resources/views/editCounter/index.html.twig
@@ -6,7 +6,8 @@
         <fieldset class="form-fieldset">
             {{ forms.wiki_input(project) }}
             {{ forms.user_input(true, false, project) }}
-
+        </fieldset>
+        <fieldset class="form-fieldset">
             <div class="checkbox">
                 <label>
                     <input type="checkbox" name="format" value="wikitext">
