+++ i18n/en.json
@@ -254,8 +254,10 @@
 	"pages-created-since-deleted": "$1 since deleted",
 	"pages-created-tooltip": "Total of all pages created, including those created as a redirect when moving a page. See the documentation for more details.",
 	"pages-deleted": "Pages deleted",
+	"pages-deleted-hover-desc": "Hover over the '$1' text to see the deletion summary.",
 	"pages-edited-total": "Pages edited (total)",
 	"pages-moved": "Pages moved",
+	"pages-recreated-desc": "Some deleted pages may have been recreated by this user or a different user. These are marked as '$1'.",
 	"pageviews": "Pageviews",
 	"pageviews-pages-created": "Pageviews of pages created",
 	"pageviews-userpage": "Userpage pageviews",
@@ -282,6 +284,7 @@
 	"reblock": "Reblock",
 	"recent-activity": "Recent activity",
 	"recent-activity-desc": "$2 {{PLURAL:$1|edit|edits}} within the last 720 days",
+	"recreated": "Recreated",
 	"redirect": "Redirect",
 	"redirects": "Redirects",
 	"redirfilter-all": "Include redirects and non-redirects",
