+++ i18n/qqq.json
@@ -253,8 +253,10 @@
 	"pages-created-since-deleted": "Text indicating the number of pages a user created that have since been deleted. $1 is the number.",
 	"pages-created-tooltip": "Text shown when hovering on a link near the 'pages created' label to explain what is excluded.",
 	"pages-deleted": "Label for pages a user has deleted.",
+	"pages-deleted-hover-desc": "In the Pages Created tool, this text indicates that you can hover the 'deleted' labels to reveal the reason why the page was deleted. $1 is the formatted message for 'deleted', which contains HTML markup.",
 	"pages-edited-total": "Label for number of total pages a user has edited.",
 	"pages-moved": "Label for pages a user has moved.",
+	"pages-recreated-desc": "Text explaining that some of the deleted pages listed may have been recreated. This is shown at the bottom of the Pages Created tool. $1 is the formatted message 'recreated', which contains HTML markup.",
 	"pageviews": "The word 'pageviews'. Used as a link to the Pageviews tool or as a label for the number of pageviews.\n{{Identical|Pageview}}",
 	"pageviews-pages-created": "Label for the pageviews of pages that were created by a user.",
 	"pageviews-userpage": "Label for the pageviews of a user's userpage.",
@@ -281,6 +283,7 @@
 	"reblock": "Name of the MediaWiki log action 'block', but describes a value that includes re-blocks.\n{{Identical|Reblock}}",
 	"recent-activity": "Label for amount of recent activity a user has taken.\n{{Identical|Recent activity}}",
 	"recent-activity-desc": "Further description for recent-activity.  $1 is the unformatted number, $2 is a formatted number",
+	"recreated": "Indicates that the page has been recreated after deletion. This message already appears alongside the text 'deleted'.",
 	"redirect": "Indicates that a wiki page redirects to another page (short word).\n{{Identical|Redirect}}",
 	"redirects": "Wiki pages that redirect to other pages (short word)\n{{Identical|Redirects}}",
 	"redirfilter-all": "Form option to show all redirect pages in the results.",
