+++ web/static/js/pages.js
@@ -5,6 +5,8 @@ $(function () {
         return;
     }
 
+    var deletionSummaries = {};
+
     setupToggleTable(window.countsByNamespace, window.pieChart, 'count', function (newData) {
         var totals = {
             count: 0,
@@ -32,5 +34,55 @@ $(function () {
             totals.redirects.toLocaleString() + " (" +
             ((totals.redirects / totals.count) * 100).toFixed(1) + "%)"
         );
+    });
+
+    $('.deleted-page').on('mouseover', function (e) {
+        var page = $(this).data('page');
+
+        var showSummary = function (summary) {
+            $(e.target).find('.tooltip-body').html(summary);
+        }
+
+        if (deletionSummaries[page] !== undefined) {
+            return showSummary(deletionSummaries[page]);
+        }
+
+        $.ajax({
+            url: wikiApi,
+            data: {
+                action: 'query',
+                list: 'logevents',
+                letitle: $(this).data('page'),
+                lestart: $(this).data('datetime').toString().slice(0, -2),
+                letype: 'delete',
+                lelimit: 1,
+                format: 'json'
+            },
+            dataType: 'jsonp'
+        }).done(function (resp) {
+            var event = resp.query.logevents[0];
+
+            // Show parsed wikitext.
+            $.ajax({
+                url: xtBaseUrl + 'api/project/parser/' + wikiDomain + '?wikitext=' + encodeURIComponent(event.comment)
+            }).done(function (markup) {
+                // Get timestamp in YYYY-MM-DD HH:MM format.
+                var timestamp = new Date(event.timestamp)
+                    .toISOString()
+                    .slice(0, 16)
+                    .replace('T', ' ');
+
+                // Add timestamp and link to admin.
+                var summary = timestamp + " (<a target='_blank' href='https://" + wikiDomain +
+                    "/wiki/User:" + event.user + "'>" + event.user + '</a>): <i>' + markup + '</i>';
+
+                deletionSummaries[page] = summary;
+                showSummary(summary);
+            }).fail(function () {
+                showSummary("<span class='text-danger'>" + $.i18n('api-error', 'Parser API') + "</span>");
+            });
+        }).fail(function () {
+            showSummary("<span class='text-danger'>" + $.i18n('api-error', 'Logging API') + "</span>");
+        });
     });
 });
